just for new merge

+++ server/src/services/data/data.service.ts
@@ -231,6 +231,11 @@ export class DataService {
     const me: User = await this.usersService.findOneById(myId);
     const friend: User = await this.usersService.findOneById(friendId);
 
+    if (!me || !friend) return { status: 500, message: 'User Not Found' };
+    const checkRequest: Request =
+      await this.requestService.findFriendRequestByUsers(me, friend);
+    if (checkRequest)
+      return { status: 500, message: 'Request is already sent' };
     const newRequest = await this.requestService.sendRequestToNewFriend(
       me,
       friend,
@@ -240,7 +245,6 @@ export class DataService {
     notification.sender = newRequest.requester;
     notification.type = NotificationType.REQUEST;
     await this.notificationsService.sendNotificationForRequest(notification);
-    console.log('salam');
     return newRequest;
   }
 
@@ -249,8 +253,14 @@ export class DataService {
     myId: number,
     friendId: number,
   ) {
-    const result = await this.addFriend(myId, friendId);
+    const me: User = await this.usersService.findOneById(myId);
+    const friend: User = await this.usersService.findOneById(friendId);
+    const checkRequest: Request =
+      await this.requestService.findFriendRequestByUsers(me, friend);
+    console.log('saidsads');
+    if (!checkRequest) return { status: 500, message: 'No Request Found' };
     await this.requestService.removeRequestById(reqId);
+    const result = await this.addFriend(myId, friendId);
     return result;
   }
 
@@ -420,11 +430,7 @@ export class DataService {
     return conversationUser;
   }
 
-  async addNewPrivateChannel(
-    file: Express.Multer.File,
-    body: CreateChannelDto,
-    myId: number,
-  ) {
+  async addNewPrivateChannel(body: CreateChannelDto, myId: number) {
     const user: User = await this.usersService.findOneById(myId);
     const owner: boolean = await this.channelService.searchForOwner(user);
     if (owner) return { message: 'U have already a channel' };
@@ -441,11 +447,7 @@ export class DataService {
     return channel;
   }
 
-  async addNewPublicChannel(
-    file: Express.Multer.File,
-    body: CreatePublicChannelDto,
-    myId: number,
-  ) {
+  async addNewPublicChannel(body: CreatePublicChannelDto, myId: number) {
     const user: User = await this.usersService.findOneById(myId);
     const owner: boolean = await this.channelService.searchForOwner(user);
     if (owner) return { message: 'U have already a channel' };
